requirements.txt
openai==2.16.0
pypdf==6.6.2
python-dotenv==1.2.1
pdf2image==1.17.0
transformers==5.0.0
torch==2.10.0
pillow==12.1.0